graphics: use the documented argument widths across the library
Audit of graphics.conf against the AmigaOS autodocs. Narrow to the
documented types: SetAPen/SetBPen/SetRast pen and SetDrMd mode (UBYTE),
BltBitMap minterm and mask, BltBitMapRastPort and BltMaskBitMapRastPort
minterm (UBYTE), BltPattern byteCnt, PolyDraw count and SetRGB4 n (WORD),
SetRGB4 r/g/b (UBYTE), FindColor maxpen (LONG). Widen AllocRaster and
AllocBitMap sizes to ULONG and WriteChunkyPixels coordinates to LONG, as
documented. Callers built for the m68k ABI leave the unused upper
register bits undefined, so the wider AROS types read garbage (see
2cb305e297).

+++ rom/graphics/bltpattern.c
@@ -63,7 +63,7 @@ static ULONG bltpattern_render(APTR bpr_data, WORD srcx, WORD srcy,
         AROS_LHA(WORD             , yMin, D1),
         AROS_LHA(WORD             , xMax, D2),
         AROS_LHA(WORD             , yMax, D3),
-        AROS_LHA(ULONG            , byteCnt, D4),
+        AROS_LHA(WORD             , byteCnt, D4),
 
 /*  LOCATION */
         struct GfxBase *, GfxBase, 52, Graphics)
